backend/requirements.txt
pdf2image
google.generativeai
python-dotenv
streamlit
